Tries

Description

This module provides an implementation along with functions for working with tries. Tries allow for a fast and scalable implementation of autocomplete.

Documentation

data Trie

Trie + autocomplete examples in a REPL environment

>>> import Tries
>>> mytrie = trieFromList [(30, "an"), (12, "to"), (25, "ted"), (20, "a"), (14, "tea")]
>>> mytrie
a 20.0 30.0
  n 30.0 -1.0
t -1.0 25.0
  o 12.0 -1.0
  e -1.0 25.0
    a 14.0 -1.0
    d 25.0 -1.0
>>> mytrie_augmented = insertWords [(5, "in"), (15, "inn")] mytrie
>>> mytrie_augmented
a 20.0 30.0
   n 30.0 -1.0
t -1.0 25.0
   o 12.0 -1.0
   e -1.0 25.0
      a 14.0 -1.0
      d 25.0 -1.0
i -1.0 15.0
   n 5.0 15.0
      n 15.0 -1.0
isInTrie "inn" mytrie
False
isInTrie "inn" mytrie_augmented
True
>>> removePrefix "inn"  mytrie_augmented
a 20.0 30.0
   n 30.0 -1.0
t -1.0 25.0
   o 12.0 -1.0
   e -1.0 25.0
      a 14.0 -1.0
      d 25.0 -1.0
i -1.0 5.0
   n 5.0 -1.0
>>> updateWeight sqrt "ted" mytrie_augmented
a 20.0 30.0
   n 30.0 -1.0
t -1.0 14.0
   o 12.0 -1.0
   e -1.0 14.0
      a 14.0 -1.0
      d 5.0 -1.0
i -1.0 15.0
   n 5.0 15.0
      n 15.0 -1.0
>>> mytrie_augmented -- Notice that this trie was never modified.
a 20.0 30.0
   n 30.0 -1.0
t -1.0 25.0
   o 12.0 -1.0
   e -1.0 25.0
      a 14.0 -1.0
      d 25.0 -1.0
i -1.0 15.0
   n 5.0 15.0
      n 15.0 -1.0
>>> autocomplete "t" mytrie 3
[(25.0,"ted"),(14.0,"tea"),(12.0,"to")]
>>> autocomplete "in" mytrie 5
[]
>>> autocomplete "in" mytrie_augmented 5
[(15.0,"inn"),(5.0,"in")]

The generic Trie data structure.

Constructors

EmptyTrie
The empty trie

PrefixNode
Non-empty tries are prefix nodes

Fields

prefix :: String

A string of length one (or 0 if we're at the root node)

weight :: Double

The weight of the prefix. This is -1 for non-words and non-negative for words

max_child_wt :: Double

The maximum weight of all children of this node

children :: HashMap String Trie

A HashMap of the children. Keys are prefixes, values are tries

rootNode :: Trie

The root node from which we can build more interesting Tries.

Higher-order functions for children of Tries

mapChildren :: (Trie -> Trie) -> Trie -> Trie

Map a function on the children of a Trie, and return the updated Trie.

filterChildren :: (Trie -> Bool) -> Trie -> Trie

Filter children of a Trie out that don't satisfy a given predicate, and return the Trie.

foldrChildren :: (Trie -> a -> a) -> a -> Trie -> a

Reduce the children of a Trie using a binary function and a starting value.

Generic functions for working with Tries

getChildrenAsList :: Trie -> [Trie]

Get the children of a Trie as a list.

getNextNode :: String -> Trie -> Trie

Given a prefix and a Trie, get the next node to search through. e.g. if prefix = "hell" and Trie has prefix of "hel" then this function will return the node with prefix "hell" (or the EmptyTrie if it doesn't exist).

getNode :: String -> Trie -> Trie

Get the node of the specified prefix in the Trie. If not in the Trie, return EmptyTrie.

isInTrie :: String -> Trie -> Bool

Check to see if a prefix is in a Trie.

getMaxWeight :: Trie -> Double

Get the maximum weight in a Trie. This includes the weight of the current node and all children.

updateMaxChildWeight :: Trie -> Trie

Update the max_child_wt of the current node (not of any children). Warning: this function assumes that all children have accurate max_child_wt values This is used to update a Trie whose max_child_wt value becomes stale once a child is deletmax_weighted (for example).

Inserting nodes into Tries

insertChild :: Trie -> Trie -> Trie

Insert a child Trie into a parent trie and update the max_child_wt of the parent.

insertChildFast :: Trie -> Trie -> Trie

Fast version of insertChild. This does NOT update max_child_wt for arbitrary insertions - this is only for adding new children to tries that are otherwise not updated. This will throw an error if the child is not valid (e.g. "the" is not a valid child of "t").

insertPrefix :: Double -> String -> Trie -> Trie

Insert a prefix and its weight into the Trie.

insertWords :: [(Double, String)] -> Trie -> Trie

Insert weighted prefixes into the Trie. This is essentially a wrapper that performs insertPrefix many times.

trieFromList :: [(Double, String)] -> Trie

Build a Trie from a list of weighted words.

Updating Tries

updateWeight :: (Double -> Double) -> String -> Trie -> Trie

Update the weight of a node using a custom update function.

insertOrUpdate :: (Double -> Double) -> String -> Trie -> Trie

Update the weight of a node using a custom update function if the node exists. If it doesn't exist, create it with 0 weight.

rescaleWeights :: (Double -> Double) -> Trie -> Trie

Rescale the weights in the Trie by an arbitrary update function.

Removing nodes from Tries

removeChild :: String -> Trie -> Trie

Remove the child (specified as a prefix) from the parent. This is primarily a helper function for removePrefix.

removePrefix :: String -> Trie -> Trie

Remove a prefix from the Trie.

pruneChildren :: Double -> Trie -> Trie

Run pruneTrie on children whose maximum weight (max(weight, max child weight)) is greater than or equal to threshold. This is effectively just a helper function for pruneTrie.

pruneTrie :: Double -> Trie -> Trie

Remove all nodes in a Trie whose weights are below a threshold.

Creating Tries from text files

textToWtdWord :: Text -> (Double, String)

Convert a correctly formed input line to a weighted word.

readTerms :: FilePath -> IO Trie

Read terms from a specified FilePath and produce a Trie.

Autocomplete

autocompletePriority :: (Double, String) -> Double

Give a priority for autocomplete. This is mainly because MinMaxQueue is a min-priority queue, so we need to flip the priority by negating the weight of the word.

mmqMinWeight :: MinMaxQueue Double (Double, String) -> Double

Get the minimum weight in the minmaxqueue The smallest weight is actually the "maximum" in the MMQ since the priorities are the negative weights

mmqInsert :: Double -> String -> MinMaxQueue Double (Double, String) -> MinMaxQueue Double (Double, String)

Insert the weighted word (the tuple (weight, word)) into q

getMatchingPrefixes :: String -> String -> Int -> Trie -> MinMaxQueue Double (Double, String) -> MinMaxQueue Double (Double, String)

Get all matching terms for a given prefix. This does the majority of the work for autocomplete.

autocomplete :: String -> Trie -> Int -> [(Double, String)]

Produce the top k words in a given Trie that match a particular prefix, sorted by weight.

slowComplete :: String -> [(Double, String)] -> Int -> [(Double, String)]

Produce the top k words in a given list of weighted words that match a particular query, sorted by weight.